+++ src/lib/index.js
@@ -1,2 +1,3 @@
 exports.stateIsUnlocked = require('./state-locks').stateIsUnlocked;
+exports.stateLock = require('./state-locks').stateLock;
 exports.stateUnlock = require('./state-locks').stateUnlock;