+++ src/lib/state-locks.js
@@ -5,8 +5,14 @@ const stateUnlock = (data, lock) => {
   }
 };
 
+const stateLock = (data, lock) => {
+  if (!data.unlocked) data.unlocked = [];
+  data.unlocked = data.unlocked.filter(l => l !== lock);
+};
+
 const stateIsUnlocked = (data, lock) =>
   (typeof data.unlocked !== 'undefined') && (data.unlocked.indexOf(lock) >= 0);
 
 exports.stateUnlock = stateUnlock;
+exports.stateLock = stateLock;
 exports.stateIsUnlocked = stateIsUnlocked;